// examples/mini-game-lab/src/fps-game-editor-adapter/types.ts
export type ProjectPersistentBinding = {
  kind: 'sceneNode';
  nodeId: string;
  rootNode: any;
};
// ...
export interface ProjectEditorPluginContext {
  scene: any | null;
  game: any | null;
}